requirements.txt
openai>=1.40.0
streamlit>=1.36.0
python-dotenv>=1.0.1
numpy>=1.26.4
PyPDF2>=3.0.1
python-docx>=1.1.2
